QPID-7739: [Java Broker] [AMQP 1.0] Change type of fields Session_1_0#_incomingWindow and Session_1_0#_outgoingWindow to UnsignedInteger

+++ broker-plugins/amqp-1-0-protocol/src/main/java/org/apache/qpid/server/protocol/v1_0/AMQPConnection_1_0Impl.java
@@ -616,7 +616,7 @@ public class AMQPConnection_1_0Impl extends AbstractAMQPConnection<AMQPConnectio
                                                           begin,
                                                           sendingChannelId,
                                                           receivingChannelId,
-                                                          getContextValue(Integer.class, AMQPConnection_1_0.CONNECTION_SESSION_CREDIT_WINDOW_SIZE));
+                                                          getContextValue(Long.class, AMQPConnection_1_0.CONNECTION_SESSION_CREDIT_WINDOW_SIZE));
                     session.create();
 
                     _receivingSessions[receivingChannelId] = session;
